Write the specification for the numeric type
Numeric attributes are numbers.

For video attributes, the values will be combined and written. The default
combiner is "average", which averages the values. Alternatives are "sum",
which sums the values, "stringify", which writes the values as a
comma-delimited concatenation of strings (and changes the output type to
"string"), and "stringifyDedup", which writes the values as a deduplicated
comma-delimited concatenation of strings (and changes the output type to
"string").