Write the specification for the datetime type
Datetime attributes are dates or times parsable by python dateutil, which
are written to the output in ISO 8601.

For video attributes, the values will be combined and written. The default
combiner is "average", which averages the UNIX times of values.
Alternatives are "stringify", which writes the values as a comma-delimited
concatenation of strings (and changes the output type to "string"), and
"stringifyDedup", which writes the values as a deduplicated
comma-delimited concatenation of strings (and changes the output type to
"string").

I, uhh, am pretty unsure of if this is a good specification, but this is
the Weka-compliant datetime way. The problem is, our two datetime
attributes are both video attributes... One is duration, which already
makes this iffy, because Weka wants a full datetime and duration doesn't
really fit that, and the other is upload date. Taking the average of the
first is good, but it would be hacky to have to make it a date and time,
and taking the average of the second doesn't seem to make sense.

In summary, we should talk about this.